Form cleanup (#352)
* Split forms into multiple files

* Removed duplicate files
