+++ members/forms/__init__.py
@@ -1,0 +1,17 @@
+from .person_form import PersonForm
+from .get_login_form import getLoginForm
+from .signup_form import signupForm
+from .volunteer_signup_form import vol_signupForm
+from .admin_signup_form import adminSignupForm
+from .activity_signup_form import ActivitySignupForm
+from .activity_invite_decline_form import ActivivtyInviteDeclineForm
+
+__all__ = [
+    ActivivtyInviteDeclineForm,
+    ActivitySignupForm,
+    PersonForm,
+    getLoginForm,
+    signupForm,
+    vol_signupForm,
+    adminSignupForm,
+]
